+++ src/Symfony/Contracts/CHANGELOG.md
@@ -5,8 +5,8 @@ CHANGELOG
 -----
 
  * added `HttpClient` namespace with contracts for implementing flexible HTTP clients
- * added `EventDispatcher\EventDispatcherInterface`
- * added `ServiceProviderInterface`
+ * added `EventDispatcherInterface` and `Event` in namespace `EventDispatcher`
+ * added `ServiceProviderInterface` in namespace `Service`
 
 1.0.0
 -----